+++ app/services/add_image.rb
@@ -1,7 +1,8 @@
 class AddImage
-  def self.call(model, public_id)
+  def self.call(model, public_id, local_id = nil)
     media_item = MediaItem.new(public_id: public_id, mediable_id: model.id,
-                               media_type: "image", mediable_type: model.class)
+                               media_type: "image", mediable_type: model.class,
+                               local_id: local_id)
     unless media_item.valid?
       model.errors.add(:mediaItem, media_item.errors.messages) && return
     end